moved wd package out of cmd dir and renamed to workingdir
