+++ cmd/create/create_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/matthewchivers/dodl/cmd/wd"
+	wd "github.com/matthewchivers/dodl/pkg/workingdir"
 	"github.com/stretchr/testify/assert"
 )
